ref: added safety port check to pre-push hook
The reason for this change is that when the pre-push hooks runs the tests, but the port at which the test-server would be running on is already in use, the tests will cause the pre-hook script to fail eventually or timeout. To avoid having to wait until this happens a simply safety check at the beginning checks whether the port is already in use or not.