ref: added click event to dialog exit button in useEfffect
The reason for this change is that the DialogContent component automatically renders an exit button at the top right corner of the dialog. However, when the dialog is controlled then the exit button has no functionality. At the same time it can not be disabled or modified, hence a click event is added dynamically through a useEffect.