feat: added Dockerfile for database
The reason for this change is that the mysql database has to have the respective tables, which cannot be creating as part of a github-action-service, because it is started before the code is checked out. In other words, when the service is created within the GitHub action it does not yet have access to the initialization script.