chore: excluded /schemas and api/version directory
The reason for this change is that both the zod functions used within the /schemas directory would require large test-suites that are not necessarily within the scope of this project. The api/version directory was ignored as it also not subject to testing as its not part of the production app.