ci: updated ignored branches in auto-link workflow
The reason for this change is that previously the workflow ignored events where a PR merged changes into the main or canary branch. Given that the workflow was intended to mainly apply to any feature PRs thus also PRs that merge changes into the canary branch, the canary ignore filter was removed.