ref: passed question-id as defaultValue if exists
The reason for this change is that when a question is edited then the "id" input's value is set to the questions-id so that the respective question will be updated instead of added as a new one. In other words, by setting passing the existing-question id to the id input, questions will be updated because the system detects that it already exists -> update.