style: changed height classes in DesktopSidebar
The reason for this change is that when the content shown exceeds the viewport / screen size a visual discrepancy appears. That means, that the a visual background difference becomes noticeable, because the desktop-sidebar previously wrapped its content in a container where the height was set to the screen-size.
By changing the height to be at least the screen size and thus not limiting the max-height this issue is resolved.