fix: resolved sidebar state changes due to re-rendering
The reason for the sidebar state to update once its mounted was due to the SidebavHoverability check. The reason for this is that the component updated the setDeviceHoverable state once mounted, changed the cached value. This means that when the sidebar was previously pinned then this information was lost as the component updated the state and thus cached / updated the state.