fix: resolved missing question answer values
The reason for this bug was that the useEffect that checks whether the type has changed to generate default values. By not considering the initial change due to the defaultValues any initial answer value was overridden.

By keeping track of the previous type actual changes are captured and thus accidental answer-value-resets avoided.