ref: reverted date property types in Kcheck schema
The reason for this change is that this way both date properties are being converted back to a date object in case they are of type string when parsed.

The main reason for removing these checks and union type definition in the first place was because the default-value generation for the closeDate produced invalid data due to the previously missing default value and use of the .transform identifier.