fix: updated storing of position to be persistent
The reason for this change is that previously the changes were stored in the form but not saved, because the onChange event of the form was not triggered. Also, the position of drag-drop items was not updated (re-rendered), because the key's of said elements were the same (not taking their current position into account).