fix: resolved failure to restore stored examination-results
The reason for this change is that the `results` property was previously initialized by default to have the same structure as the questions of the respective knowledgeCheck. However, this initialization did not take in account that the `results` property may already be defined, thus always resetting it to the default-structure.