fix: resolved layout shifts due to size-auto class
The reason for this change is that this class caused a layout shift when the /checks page was first loaded. This was because the size chosen was larger than the actual size of the icon, thereby causing a layout shift. By removing the `size-auto` class this issue has been resolved