ref: discarded cached value if ids do not match
The reason for this change is that previously knowledgeChecks were sometimes rendered incorrectly due to cached data. More specifically if an user started to edit or create a knowledgeCheck, which was cached. Then this cache could lead to overriding the initialProps of the store, when a user starts editing a different check. This means that the previously edited check would be displayed even though the user expected a different check to be displayed.