fix: resolved defaultValue declaration for closeDate
The reason for this change is that when the defaultDate was internally converted to a string instead of Date, then this usage throws an error because the toDateString() function does not exist on string type. Instead, the closeDate is always parsed as a Date ensuring that a string is converted back to a Date. In case the closeDate is not set the current date will be used instead.