fix: resolved component parallelization test-script
The reason for this change is that the execution of component tests does not require the production server that was started in the beginning by the test-script. In fact, having this production server still up and running from the e2e tests, causes an internal error due to the port being already in use.

By terminating the production server that was used for running the e2e after they are finished, eliminates the port-in use error that sometimes occurred when executing the component tests in parallel.