fix: resolved unwrapping effects in stripZodDefault
The reason for this change is that schemas that were wrapped in effects where not unwrapped. This meant that e.g the KnowledgeCheck schema that was wrapped in a Refinement had all its default values. Hence, parsing an object that had undefined values lead to the fact that if these values had default-values that these values were used instead of throwing an zod-error for required fields.