fix: re-applied zod-effects to stripped zod schema
The reason for this change is that previously any effect that was wrapping another schema was removed when the stripZodDefault utility function was applied. Thats because it unwrapped the types to remove default values. However, it did not re-apply the stripped effects, which meant that said effects were lost. This means that using the validate functions did not consider these effects as they were removed.