fix: added required id property to baseValues
The reason for this change is that by removing the default-value in the question-schema to achieve unique values, submitting the form when creating a new check was unsuccessful. This was because the id was not set and thus the form could not be submitted. By adding an id to the baseValues the required-id constraint is satisfied.