ref: improved Line animation color display
The reason for this change is that when the progress-line was animated from the left to the right, then the completion color (blue) was set immediately. This meant that the animation was almost invisible because it used the same color. Now the line that is to be animated will animate its color transition with a duration of 1000 when animation from the left to the right. This means that the sliding animation will be done before line is fully colored with the completion-color (blue).