ref: simplified useFilterStages logic
The reason for this change is that the amount of stages displayed in the condensed view / variant should always be the same => max 3. Hence, the amount of conditions can be reduced as the first and last stage is always displayed. Only the stage in-between these boundaries depends on what stage is currently shown.