style: updated button cursor styles when btn disabled
The reason for this change is that previously the button lost its interactiveness when disabled. Meaning that it did not show a respective cursor. By changing the hover and active styles to only appear when the button is enabled, thus not disabled. The class "pointer-events-none" was removed, which previously removed these styles, but also removed the cursor-not-allowed when hovering over the disabled button.