ref: renamed CreateCheckStore components and types
The reason for this change is that by renaming all components, types and functions from `CreateCheck[Store, Provider, ...]` to Check[Store, Provider, ...] makes it more generic and thereby more reusable. Furthermore, the previous naming pattern of all related elements of the "CreateCheckStore" were not only used to keep track of a created check, but was also used to when editing a check. Hence, the naming was updated.