ref: added debug log statement when no cacheKey entry was found
The reason for this change is that in case a user saved a new knowledgeCheck, but defined a custom cacheKey in the CheckProvider then the SaveCheckButton would not delete the respective cache when the cacheKey is not also provided to the SaveCheckButton. To inform users in such an edge case about not finding a respective cache to delete, debugging is easier.