ref: redefined type declarations for `zustand` library
The reason for this change is that the previously declared type definitions (`CreateStoreType`) were a bit obscure. Additionally, they provided the caching related properties to every store-function even though if the store function is not meant to be cached.

By introducing these new type declarations the previous shortcomings should be resolved.