ref: updated useCacheCreateStore types
The reason for this change is that the new `zustand` related types require that the useCacheCreateStore function now also accepts the Store as its generic type.

This way the type-definition for the `createStoreFunc` property can be set accordingly. This also means that the State type that was previously for the type-definitions of the remaining properties has to be retrieved from the generic Store type.