feat: created `createZustandStore` utility function
The reason for this new utility function is that it de-duplicates the instantiations of new stores that are cached. In other words, previously every store that was cached used the same boilerplate code of defining the `modify` function that is used instead of the set function to cache changes.

By using this new createZustandStore function the caching-setup is done under the hood, reducing the overall code duplication.