ref: updated StoreCachingOptions props
The reason for this change is that when caching is enabled then a cacheKey must be set. This way the provided cacheKey can then be used both for the cache-retrieval and for caching, as it can be passed along.

The same is true for the debounceTime, which was moved from the useCacheStoreUpdate declaration into the StoreCachingOptions type, so that it can be passed directly as a store-option.