ref: updated MultiStage store and provider
The reason for this change is that previously with the old store and provider implementations the initialState was of type Partial<MultiStageState>. This meant that the stage property could be left out from the initial props. However, the previous type definition of the `ZustandStore` type required the initialProps to be either undefined or satisfy the StoreState. This meant that when the StoreState had no optional arguments they had to be provided in the initialState prop.

Now, by overriding the initialState type, its behavior is the same as before.