ref: removed unused config property from Sidebar state
The reason for the removal of the config property is simply that is only used once. That and the fact that including it means that it has to be removed before caching and manually appended when restored, means that it does not make sense to keep it in the store-state.