feat: applied react-hook-form validation to RenderPracticeQuestion
This way both client-side and server-side errors will be displayed to the users. At the same time, users first have to check their answers before they can continue to the next question. Which allows the webpage to provide them with the solution, after which they can decide to continue to the next question.