ref: updated Practice schema for multiple-choice answers
The reason for this change is that by default the selection-array will have a length equal to the amount of answer-options. When an answer is not selected, the value at the question-index is empty / undefined, because it is not selected. By updating the schema to be an array of optional strings, this behavior no longer causes an form error when submitting.