ref: replaced RenderQuestionType with conditional rendering
The reason for this change is that when using the RenderQuestionType to render the different inputs based on the question-type an error occurs. More specifically, registering inputs for both single- and multiple- choice questions worked. However, registering the inputs for both the dragDrop and open-question caused an error:

Cannot update a component (RenderPracticeQuestion) while rendering a different component (RenderQuestionType). To locate the bad setState() call inside RenderQuestionType.