ref: added custom onChange handler to multiple-choice answers
The reason for this change is that previously, when no an answers was not selected the value at the given index was of type boolean ("false"), which violates the schema definition. By changing the onChange handler to use the answer.name instead, the selection array either includes the answer.name or undefined.