ref: updated practice-schema multiple-choice selection type
The reason for this change is that by using an input of type checkbox the value of the respective field is either the answer-name or "false" if not selected. The previous schema-definition that required the selection to be either optional or of type string was thereby never satisfied unless all options were selected. By updating the schema-type to accept both strings and booleans and transforming the booleans (false) to null.