ref: changed multiple-choice input registration
The reason for this change is that previously, the registration of the multiple-choice selection-input caused the form-state to get updated twice, when the input first changed. Thats because these inputs were effectively registered on their first-change thus causing the form-state to update.

By replacing the custom onChange handler with the basic register function and the updates to the practice-schema the behavior remains the same, while removing the unnecessary form-update.