ref: updated .default operand to use functions in QuestionSchema
The reason for this change is that this way the default-values for each type of question are not the same for each instantiation. Previously the returned array was the same for every instantiation. By wrapping the default-value in a function the output is re-generated on every instance making the default-values unique.