ref: separated question-schemas and fixed utility functions
The reason for this change is that previously the utility functions for each type of question were essentially instantiating a generic question. This means that the type of question that was generated was chosen by random. By separating the schema definitions the respective utility functions were updated to instantiate the respective questions properly.