chore: installed server-only dependency
With the help of this dependency modules can be marked as server- or client-only. This way an error will be thrown within this module when it is used within a different environment. Thus, when 'server-only' is used and the module is used on the client an error will be thrown.