fix: updated multiple-choice practice schema
The reason for this change is that when the multiple-choice question not rendered first, the selection property does not exist. Thus, selecting individual options adds them to selection array. However, when an option in between is left out, the value on that answer's index is `undefined`, thus violating the previous schema. Hence, users were unable to check their answers in this situation.

To prevent this the selection property can now be either of type uuid, null, false, undefined, which is transformed to uuid or null.