ref: changed return-type based on provided question-type
This way the utility function returned by the usePracticeFeedback hook returns different arguments depending on the question type.

This means that for ChoiceQuestions additional utility functions like isCorrectlySelected are available. While these functions are not available for other types of questions like "open-question".

Hence, the type-declarations of the getFeedbackEvaluation function were updated extensively.