ci: updated db-docker-build to use dynamic tags
The reason for this change is that when the database-schema is first modified a docker-image will be created for that specific commit. When the changes are eventually merged into the canary branch a canary-version will be published and eventually when the changes are merged into the main branch a stable image will be published.

This users can decide upon which docker-image to use based on stability.