ci: ignored main and canary branch in db-docker-image workflow
The reason for this change is that this way the workflow only creates a new database-docker image for a feature branch. The reason for this is that when the a docker-image for the main or canary branch is created the most-recent tag should be used. This then depends on the semantic-release workflow on whether or not a new release and thereby new tag is created.