ref: simplified store share_token query
The reason for this change is that the share-key is now marked as unique within the database. This means that when a share-token would be used twice an error will be thrown by the update statement. This means that we must no longer check before updating whether a key is already in use, as it is done by the database.