feat: introduced compose-init-script
This script takes in two arguments. The first one being the directory where the drizzle-migrations are stored, the second one is the output-file. It essentially aggregates the individual drizzle-migration steps into one init.sql file. This init.sql will then be  used to setup the database when building the database docker-image, like before.