chore: added default value for options.type
Previously the value of the options.type property was undefined when not provided, this meant that the if-condition that updated the selection was never used. By simply overriding the type property at the beginning of the command to the question.type this issue is prevented.

The reason why the options.type prop exists in the first place is to type assert the option based on the question type.