chore: changed simulate drag-drop answering to keep order
Previously the order at which drag-drop options were positioned was sorted again, which meant that they were always placed in the correct order. By simply using the order in which they were provided, thus not sorting them based on their position again, the answering of drag-drop questions works as expected.