chore: removed test-data cleanup from cypress.config
The reason why the cleanup of test-data was moved away from the "after:run" hook inside the cypress.config is that when tests are executed in parallel, each set of suites finished within a thread will trigger the cleanup. This means that test-data is cleaned while other tests may still rely on said data, because they are executed simultaneously.