fix: resolved adding question when edited
The reason for this bug was that when the type was changed the id was reset as well causing it to get a new id. Hence, when the question was updated it was identified as a new question because the id did not yet exist. By simply re-using the id when resetting form inputs this bug should be resolved.