fix: resolved form-fields cache validation in CreateQuDialog
Previously, when the user began editing a question, but then closed the dialog without resetting the fields, e.g. by clicking outside the dialog. Then, started to edit another question the field-values of the other question were displayed instead of the actual question that user wanted to edit.

Hence, when the dialog is opened it will not check whether the cached form-values match the respective question based on its id. When that's not the case the form will be reset. At the same time when the form is reset the dialog will now differentiate between editing and creating, meaning that when a check is edited, resetting will reset the form based on the to-edit-question.