ref: updated reseting of CreateQuDialog when type is changed
This way the user-inputs of the answers are preserved when the user switches between compatible question-types. For example, when switching between Choice and DragDrop questions the answer-option values will be preserved. However, when the user switches to a non compatible question like open-question then these inputs will be lost, unless the user is editing a question, then these values will be re-used when switching the original-edit values. Otherwise, the default-values of any other compatible question that is selected after the "open-question" question are re-used again.