chore: increased anonymous avatar get timeout to 10 seconds
The reason for this change is that in some occasions the time it takes until the user avatar is displayed exceeds the default timeout of 4 seconds. This then creates the impression of flaky test outcomes. By increasing the timeout to 10 seconds the probability of slow avatar's causing tests to fail should be almost 0.