ref: updated social login test-cases
These test-cases were updated because the google-sign in request is now blocked by google thus causing the test-case to fail. To avoid the same issue with the GitHub provider both tests were updated to verify that the sign-in request is sent to the /api endpoint instead of verifying the user is redirected.

The reason for the removal of this test is that with the new cypress version the redirect to sign-in with google is blocked by google. At the same time the redirect url cannot be