ci: removed caching of instrumented-build
The reason for the removal of caching the instrumented-build is that with the new `next` version the build time is drastically reduced from around 2 minutes down to 30 seconds in the pipeline. Hence, caching the build takes longer than building the app separately in each subsequent step. Caching the build takes around 1 minute, while simply building it takes around 30 seconds.