ref: added default type prop to AuthButton
The reason for this change is that this way the default button type of `submit` is changed to `button`. This way authentication buttons will submit a form they may be part of when clicked.

This was also the reason why clicking the dex-provider button was not necessarily successful when clicked the first, because the submit-event caused the page to reload before the authentication request could redirect the user to the respective provider page.