ci: introduced matrix to semantic-versioning docker-build job
This way the docker-image for each platform is build and pushed within a separate job. This should improve the overall performance of the build-job and should require less time until both images are build and pushed.

Previously, the download and build step of the second platform took almost 10x the time needed for the first download and build steps.