ref: added `_position` to drizzle database schema
A new `_position` property is introduced to the `db_question` and `db_answer` tables to retrieve entries in the same order they were inserted. The reason for this is that the createdAt property is identical when inserting a small number of questions / answers, thus the order does not necessarily match the one from the KnowledgeCheck object. For this reason the `_position` property is assigned the index of the question so that they can be sorted based on the index aka. _position.