ref: updated flaky assertion in `SessionAStorageCache.cy.tsx`
The reason for this change is that previously when the session-storage entries were invalided and thus empty for the domain, the assertion failed. This was because the session-storage was completely empty in some cases, thus accessing the domain-storage threw an error within cypress, which caused the assertion to fail.

The updated assertion now handles both cases were the sessionStorage is empty, thus not including the respective store-entry and where the storage is not empty by checking if the store property is not part of the storage.