ref: added single-item array logic to `shuffleArray`
By introducing a simple logic that checks whether the given array has a length of 0 or 1, no error will be thrown in such a case, but instead an warning message is shown.  The reason for this is that an array with such a length cannot be shuffled meaningfully, so the original array is simply returned.