ci: replaced static `.env` file declaration with dynamic approach
This way the environment-variables must not be manually updated whenever they change within the workflow. Instead the repository `secrets` and `variables` that are used indirectly anyway are simply printed into the `.env` file. This means that whenever the environment-variables change they must only be updated within the repository `secrets` and `variables` and they are automatically updated in the `semantic-versioning` workflow.