ci: replaced static workflow `env` declaration to be dynamic
This way the environment-variables must not be manually updated whenever they change within the workflow. Instead the repository `secrets` and `variables` that are used indirectly anyway are simply printed into the `$GITHUB_ENV` placeholder. This means that whenever the environment-variables change they must only be updated within the repository `secrets` and `variables` and they are automatically updated.