ref: called `startExamination` on `ExamQuestionWrapper` render
This way the exam attempt is cached as soon as the `ExamQuestionWrapper` component is mounted. This means that the state will not be re-created when the page is refreshed, because the cache will be used instead. This also includes the state property `startedAt`. Hence, users cannot reset their `remainingTime` by refreshing the page, as the value is cached and restored by the `ExaminationStore`.