ref: replaced `PracticeStore` usage with Link tag
Instead of using the store to update the selected category, users click on a respective `Link` to open the practice page with a category set as a searchParam. Thereby, prefetching the practice page and at the same time not relying on the `PracticeStore` and thus not needing a shared `PracticeProvider`.