ref: simplified `insert` category handler
This way the category `insert` handler no longer handles the duplicate entry event, but instead throws an error when a duplicate category is inserted. The reason for this change is that the `.onDuplicateKeyUpdate` utility function does not return the existing category-id when a duplicateKey error is suppressed. Hence, the error is essentially unhandled and passed along.