fix: updated rule suggestions on per-owner basis
The reason for this change is that previously, suggestions were overridden by eslint or the IDE when there were multiple suggestions for the same node/line. This caused the correct suggestion to appear in the suggestions-popup but, their underlying fixer was overridden with the latest suggestion for that respective attribute/line.

Now, suggestions are made per "owner" meaning per actual line / argument of utility functions. This eliminates the issue.