style: added light + focus styles to correctness input
This way users are shown visual feedback when the correctness input is focussed. At the same time the label that wraps around the correctness input is no longer accessible via tab. The reason for this is that users can still access the input via tabbing even though its hidden, but when selected the label shows its focus state.