ref: updated `evaluateClassname` to better detect variable-names
Previously (custom) variables names that contained dashes (-) were not recognized because the className is split by the dashes and previously only the first argument was used to identify whether it's a color or not. Now the hasColor check also verifies whether a custom variable that includes several dashes is used.