style: updated styles in `Select` component
The `Select` component now features modified classes for the `CommandItem` component that improves the display of hovered items within the popover-content. By default, a CommandItem stays selected once hovered, until the popover-content is closed again. Hence, the selection styles were disabled by default except on-hover. This way items can be hovered and are highlighted as such, but when the select popover-content loses focus / is no longer hovered the selection will no longer be highlighted.