ref: improved UX by refreshing page, scroll to top
Whenever a user creates a new share-token for a check that previously did not exist then the underlying check is updated and thus moved to top (sorted by last-update). Thus, the `ShareKnowledgeCheckButton` now refreshes the page when a new share-token is inserted and thus the respective check is updated. So that updated checks are fetched again with the newly introduced `share-token`. At the same time, the button will scroll to the top of the window, to bring the updated check back into view (which should be the top-most check, because it was the last one that was updated).